requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-WTF==1.2.2
Flask-Cors==5.0.0
gunicorn==23.0.0
gevent==24.11.1
python-dotenv==1.0.1
werkzeug==3.1.3
